Change of strategy. from three version of the app (txt, CSV or mongoDB) merging all-in-one - changin in setup and app_config to accomodate the merge - removing file_class non-class functions - created csv_module to  inherit stray functions from file_class - several changes in web_module to accomodate new requirements - update yml files to new requirements
